memory-service/pyproject.toml
[project]
name = "arc-memory"
version = "0.1.0"
description = "Local RAG memory layer for ARC-Hub — semantic search over 1000+ past AI conversations"
requires-python = ">=3.11"
dependencies = [
    "lancedb>=0.6.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "watchdog>=4.0.0",
    "python-frontmatter>=1.1.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.7.0",
    "pyarrow>=15.0.0",
]

[project.scripts]
arc-ingest = "ingestion.run_ingest:main"
arc-serve = "mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "retrieval", "mcp_server", "prompting"]
